connect backend to frontend

+++ src/main/java/com/revature/eCommerce/resposity/CartRepository.java
@@ -7,7 +7,8 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface CartRepository extends JpaRepository<ShoppingCart, Integer> {
+public interface CartRepository extends JpaRepository<ShoppingCart, Long> {
     //Find SHopping cart base on User ID
     Optional<ShoppingCart> findByUser_userId(Integer userId);
+
 }
